Update Document
Mobile devices are getting more advanced every couple of years. Back in the early 2,000's we had black berry's and now we have mini computer in our pockets. For example "he original BlackBerry devices, the RIM 850 and 857, used the DataTAC network. In 2002, the more commonly known convergent smartphone BlackBerry was released, which supports push email, mobile telephone, text messaging, Internet faxing, Web browsing and other wireless information services" (https://en.wikipedia.org/wiki/BlackBerry#:~:text=The%20original%20BlackBerry%20devices%2C%20the,and%20other%20wireless%20information%20services). According to that information it was like the new of technology. Everyone was really hyped and excited because you could text on  it really fast, better then the flip phone and you could also go on emails and access the main GUI faster. A couple of years later the new iPhone came out in 2007. The new iPhone was the iPhone 3gs. People went crazy over this phone because it was touch screen, and everything you need was in one click of the home button. The iPhone has really developed since it was ever first made. For example "iPhone with 3G networking that is twice as fast* as the first generation iPhone, built-in GPS for expanded location based mobile services, and iPhone 2.0 software which includes support for Microsoft Exchange ActiveSync and runs the hundreds of third party applications already built with the recently released iPhone SDK. In the US the new iPhone 3G is priced at a stunning $199 for the 8GB model, and just $299 for the 16GB model." (https://www.apple.com/newsroom/2008/06/09Apple-Introduces-the-New-iPhone-3G/). That is basically saying that the iPhone was the start of the future for phones. Just in 2008 we had GPS tracking and network calling and texting. Even though when it came out it has only $200 which isn't that bad for a iPhone. Now they range to $600-$1,000. Every since that they phones have been growing and becoming more smarter. You can literally talk to a phone and it will answer you back and respond to you. It takes in commands that you want to give it. Therefore the future is already here. We have mini computers in our pockets with such high processing power and such quick response time and the feel of it being worth the money.